Add function to extract access token from JSON cache
Access tokens can be used for the next 3600 seconds after generation and
after this time the key is going to be invalidated. To reduce the amount
of HTTP requests sent to the API service the library will cache the data
for as long as necessary in the disk and will use the token defined in
it until the expiration time arrives, then will request a new token.