Fix HTTP request execution and JSON parsing
Reddit API requires you to set a custom User-Agent for all the HTTP
requests sent to their service, by default the library will set a basic
string with the version number and URL of the repository containing the
code and will allow the developer to set a different string. Notice that
errors during the execution of the HTTP requests and parsing of the JSON
data is passed all the way back to the origin of the function call, you
will be asked to handle these errors in your application as exceptions.