Modify HTTP request functions for string and json
This commit separates the functions to retrieve the data from the API
service when the data is expected to be a raw string or a JSON object in
which case will be automatically decoded but in the case of raw strings
will be simply returned to the parent caller. This allows to interact
with API endpoints that do not return JSON objects but image blobs, HTML
or simply raw strings.