Modify data type returned from non-JSON HTTP responses
Go developers prefer to manipulate data as arrays of bytes instead of
strings, so with this commit I am converting the responses from the HTTP
requests to []byte for manageability. Notice that the API endpoints with
JSON responses are still being automatically decoded and any error
resulting from the parsing operation will be returned to the parent
caller.