Add function to generate and retrieve a new captcha
Request a CAPTCHA image given an iden. An iden is given as the captcha
field with a BAD_CAPTCHA error, you should use this endpoint if you get
a BAD_CAPTCHA error response. Responds with a 120x50 image/png which
should be displayed to the user. The user's response to the CAPTCHA
should be sent as captcha along with your request. To request a new
CAPTCHA, use /api/new_captcha. Below is an example of how to use it:

file, _ := os.Create("foobar.png")
data, _ := redditapi.CaptchaIden("12345")
io.Copy(file, data)