Add function to post a new comment in global state
Submit a new comment or reply to a message. The parameter "parent" is
the fullname of the thing being replied to. The parameter "text" should
be the raw markdown body of the comment or message. To start a new
message thread, use /api/compose. The value of "parent" changes the kind
of object created by this request:

- fullname (link): a top-level comment in that Link's thread.
- fullname (comment): a comment reply to that comment.
- fullname (message): a message reply to that message.